Agent: Rename HTTPRequestsFacade to HTTPClient
