+++ monkey/infection_monkey/island_api_client/http_island_api_client.py
@@ -20,7 +20,7 @@ from . import (
     IslandAPIRequestError,
     IslandAPIRequestFailedError,
 )
-from .http_requests_facade import HTTPRequestsFacade, handle_island_errors
+from .http_client import HTTPClient, handle_island_errors
 
 logger = logging.getLogger(__name__)
 
@@ -46,7 +46,7 @@ class HTTPIslandAPIClient(IIslandAPIClient):
         agent_event_serializer_registry: AgentEventSerializerRegistry,
     ):
         self._agent_event_serializer_registry = agent_event_serializer_registry
-        self.request_facade = HTTPRequestsFacade("")
+        self.http_client = HTTPClient("")
 
     @handle_island_errors
     def connect(
@@ -54,16 +54,16 @@ class HTTPIslandAPIClient(IIslandAPIClient):
         island_server: SocketAddress,
     ):
         api_url = f"https://{island_server}/api"
-        requests_facade = HTTPRequestsFacade(api_url)
+        requests_facade = HTTPClient(api_url)
         requests_facade.get(  # noqa: DUO123 type: ignore [attr-defined]
             endpoint="",
             params={"action": "is-up"},
         )
-        self.request_facade = requests_facade
+        self.http_client = requests_facade
 
     @handle_island_errors
     def send_log(self, agent_id: AgentID, log_contents: str):
-        self.request_facade.put(
+        self.http_client.put(
             f"agent-logs/{agent_id}",
             MEDIUM_REQUEST_TIMEOUT,
             log_contents,
@@ -72,18 +72,18 @@ class HTTPIslandAPIClient(IIslandAPIClient):
     @handle_island_errors
     def get_agent_binary(self, operating_system: OperatingSystem) -> bytes:
         os_name = operating_system.value
-        response = self.request_facade.get(f"agent-binaries/{os_name}", MEDIUM_REQUEST_TIMEOUT)
+        response = self.http_client.get(f"agent-binaries/{os_name}", MEDIUM_REQUEST_TIMEOUT)
         return response.content
 
     @handle_island_errors
     def send_events(self, events: Sequence[AbstractAgentEvent]):
-        self.request_facade.post(
+        self.http_client.post(
             "agent-events", MEDIUM_REQUEST_TIMEOUT, self._serialize_events(events)
         )
 
     @handle_island_errors
     def register_agent(self, agent_registration_data: AgentRegistrationData):
-        self.request_facade.post(
+        self.http_client.post(
             "agents",
             SHORT_REQUEST_TIMEOUT,
             agent_registration_data.dict(simplify=True),
@@ -92,7 +92,7 @@ class HTTPIslandAPIClient(IIslandAPIClient):
     @handle_island_errors
     @convert_json_error_to_island_api_error
     def get_config(self) -> AgentConfiguration:
-        response = self.request_facade.get("agent-configuration", SHORT_REQUEST_TIMEOUT)
+        response = self.http_client.get("agent-configuration", SHORT_REQUEST_TIMEOUT)
 
         config_dict = response.json()
         logger.debug(f"Received configuration:\n{pformat(config_dict)}")
@@ -102,7 +102,7 @@ class HTTPIslandAPIClient(IIslandAPIClient):
     @handle_island_errors
     @convert_json_error_to_island_api_error
     def get_credentials_for_propagation(self) -> Sequence[Credentials]:
-        response = self.request_facade.get("propagation-credentials", SHORT_REQUEST_TIMEOUT)
+        response = self.http_client.get("propagation-credentials", SHORT_REQUEST_TIMEOUT)
 
         return [Credentials(**credentials) for credentials in response.json()]
 
@@ -121,13 +121,13 @@ class HTTPIslandAPIClient(IIslandAPIClient):
     @handle_island_errors
     @convert_json_error_to_island_api_error
     def get_agent_signals(self, agent_id: str) -> AgentSignals:
-        response = self.request_facade.get(f"agent-signals/{agent_id}", SHORT_REQUEST_TIMEOUT)
+        response = self.http_client.get(f"agent-signals/{agent_id}", SHORT_REQUEST_TIMEOUT)
 
         return AgentSignals(**response.json())
 
     @handle_island_errors
     def get_agent_plugin(self, plugin_type: PluginType, plugin_name: str) -> bytes:
-        response = self.request_facade.get(
+        response = self.http_client.get(
             f"/api/agent-plugins/{plugin_type.value}/{plugin_name}", MEDIUM_REQUEST_TIMEOUT
         )
